REF: organize Timestamp constructor tests
