+++ pandas/tests/scalar/timestamp/test_timestamp.py
@@ -262,6 +262,14 @@ class TestTimestampProperties:
 
 
 class TestTimestamp:
+    @pytest.mark.parametrize("tz", [None, pytz.timezone("US/Pacific")])
+    def test_disallow_setting_tz(self, tz):
+        # GH#3746
+        ts = Timestamp("2010")
+        msg = "Cannot directly set timezone"
+        with pytest.raises(AttributeError, match=msg):
+            ts.tz = tz
+
     def test_default_to_stdlib_utc(self):
         assert Timestamp.utcnow().tz is timezone.utc
         assert Timestamp.now("UTC").tz is timezone.utc
@@ -378,52 +386,6 @@ class TestTimestamp:
         assert stamp.day == 21
         assert stamp.microsecond == 145224
         assert stamp.nanosecond == 192
-
-    @pytest.mark.parametrize(
-        "value, check_kwargs",
-        [
-            [946688461000000000, {}],
-            [946688461000000000 / 1000, {"unit": "us"}],
-            [946688461000000000 / 1_000_000, {"unit": "ms"}],
-            [946688461000000000 / 1_000_000_000, {"unit": "s"}],
-            [10957, {"unit": "D", "h": 0}],
-            [
-                (946688461000000000 + 500000) / 1000000000,
-                {"unit": "s", "us": 499, "ns": 964},
-            ],
-            [
-                (946688461000000000 + 500000000) / 1000000000,
-                {"unit": "s", "us": 500000},
-            ],
-            [(946688461000000000 + 500000) / 1000000, {"unit": "ms", "us": 500}],
-            [(946688461000000000 + 500000) / 1000, {"unit": "us", "us": 500}],
-            [(946688461000000000 + 500000000) / 1000000, {"unit": "ms", "us": 500000}],
-            [946688461000000000 / 1000.0 + 5, {"unit": "us", "us": 5}],
-            [946688461000000000 / 1000.0 + 5000, {"unit": "us", "us": 5000}],
-            [946688461000000000 / 1000000.0 + 0.5, {"unit": "ms", "us": 500}],
-            [946688461000000000 / 1000000.0 + 0.005, {"unit": "ms", "us": 5, "ns": 5}],
-            [946688461000000000 / 1000000000.0 + 0.5, {"unit": "s", "us": 500000}],
-            [10957 + 0.5, {"unit": "D", "h": 12}],
-        ],
-    )
-    def test_unit(self, value, check_kwargs):
-        def check(value, unit=None, h=1, s=1, us=0, ns=0):
-            stamp = Timestamp(value, unit=unit)
-            assert stamp.year == 2000
-            assert stamp.month == 1
-            assert stamp.day == 1
-            assert stamp.hour == h
-            if unit != "D":
-                assert stamp.minute == 1
-                assert stamp.second == s
-                assert stamp.microsecond == us
-            else:
-                assert stamp.minute == 0
-                assert stamp.second == 0
-                assert stamp.microsecond == 0
-            assert stamp.nanosecond == ns
-
-        check(value, **check_kwargs)
 
     def test_roundtrip(self):
         # test value to string and back conversions
